Resolved merge conflict between between develop and database-login, enabled database authentication.
With this commit we are finally doing database authentication,
username is queried from member table and password is compared to the
password hash in that same table. Some mock data is in the database
unstable, so use the environment variable ENVIRONMENT=unstable to get
working login. Also required to set variable PG_PASSWORD to get
database connection, this of course requires you to have a working SSH
tunnel to the server as well, refer to the email by Kimmo for
instructions.

Conflicts:
	routes.js