Make backend wait for database to be up in docker-compose.
Added a healtcheck for database to see if it is initiliazed. Backend
waits for database healtcheck to pass before starting.

When running 'docker-compose -f docker-compose.dev.yaml up' the backend
service crashes because it does not wait for the database service to be
up. This wouldn't be a problem otherwise but at least on my system
the output of the backend service is not shown after the crash.