Bug fix for the shooting track visitor number stats
In the tablet view, when incrementing the number of visitors, if
the status of the range officer wasn't changed once (or multiple times)
the request for updating the visitor count for a range failed with
status 500 and thus no updates would go through.

Issue was that the state of the supervisor was initialized incorrectly
to be in Finnish when the backend only accepts either 'present',
'absent' or 'closed' when updating.

Fixed by initializing the state with the data that was fetched from the
backend in the first place.